requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
aiosqlite==0.22.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asttokens==3.0.1
attrs==26.1.0
banks==2.4.2
beautifulsoup4==4.14.3
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.3
colorama==0.4.6
comm==0.2.3
cryptography==47.0.0
dataclasses-json==0.6.7
debugpy==1.8.20
decorator==5.2.1
Deprecated==1.3.1
dirtyjson==1.0.8
distro==1.9.0
executing==2.2.1
fastapi==0.136.1
filetype==1.2.0
frozenlist==1.8.0
fsspec==2026.4.0
google-ai-generativelanguage==0.6.15
google-api-core==2.30.3
google-api-python-client==2.195.0
google-auth==2.50.0
google-auth-httplib2==0.3.1
google-genai==1.74.0
google-generativeai==0.8.6
googleapis-common-protos==1.74.0
greenlet==3.5.0
griffe==2.0.2
griffecli==2.0.2
griffelib==2.0.2
groq==1.2.0
grpcio==1.80.0
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
hyperframe==6.1.0
idna==3.13
ipykernel==7.2.0
ipython==9.13.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.14.0
joblib==1.5.3
jupyter_client==8.8.0
jupyter_core==5.9.1
llama-index==0.14.21
llama-index-core==0.14.21
llama-index-embeddings-gemini==0.4.2
llama-index-embeddings-openai==0.6.0
llama-index-instrumentation==0.5.0
llama-index-llms-gemini==0.6.2
llama-index-llms-openai==0.7.7
llama-index-vector-stores-qdrant==0.10.0
llama-index-workflows==2.20.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib-inline==0.2.1
multidict==6.7.1
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.6.1
nltk==3.9.4
numpy==2.4.4
openai==2.33.0
packaging==26.2
parso==0.8.6
pillow==10.4.0
platformdirs==4.9.6
portalocker==3.2.0
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.27.2
protobuf==5.29.6
psutil==7.2.2
pure_eval==0.2.3
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.3
pydantic_core==2.46.3
Pygments==2.20.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pywin32==311
PyYAML==6.0.3
pyzmq==27.1.0
qdrant-client==1.17.1
regex==2026.4.4
requests==2.33.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.49
stack-data==0.6.3
starlette==1.0.0
tenacity==9.1.4
tiktoken==0.12.0
tinytag==2.2.1
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.46.0
wcwidth==0.6.0
websockets==16.0
wrapt==2.1.2
yarl==1.23.0
